Refactor library to remove dependency on Position
The position of devices was previously managed by the Pica
context instance and updated by pica commands such as
SetPosition/InitUciDevice.

This change introduces the RangingEstimator trait, and adds
an instance of ranging estimator to the parameters of the
Pica::new() function.

Instead of relying on device positions, ranging measurements
are now produced by the owned instance of RangingEstimator.

The http server _which receives HTTP requests to update
device positions_ now caches the device positions instead of the
Pica context, and implements a RangingEstimator to pass
to the Pica context.