Reorganize definitions in uci_packets.pdl
Instead of deriving packets based on the message type
first then the gid + oid, which prevents proper typing
of the oid; packets are written to derive from the gid
first, then mt + oid.

This way, the oid can have the appropriate enum type
and commands do not have to be written with hardcoded
oids.